refactor: Improve aircraft image fetching with new mapping system
